* Merge SVN revision r1240 from branch devel into stable.
git-svn-id: svn://svn.code.sf.net/p/colinux/code/branches/stable@1241 688f6386-cf32-0410-9ead-a596f0484e40

+++ src/colinux/os/kernel/alloc.h
@@ -22,13 +22,13 @@ extern co_rc_t co_os_get_page(struct co_manager *manager, co_pfn_t *pfn);
 extern void *co_os_map(struct co_manager *manager, co_pfn_t pfn);
 extern void co_os_unmap(struct co_manager *manager, void *ptr, co_pfn_t pfn);
 extern void co_os_put_page(struct co_manager *manager, co_pfn_t pfn);
-extern void *co_os_alloc_pages(unsigned long pages);
-extern void co_os_free_pages(void *ptr, unsigned long pages);
+extern void *co_os_alloc_pages(unsigned int pages);
+extern void co_os_free_pages(void *ptr, unsigned int pages);
 
 /*
  * Interfaces for mapping physical memory in userspace.
  */
-extern co_rc_t co_os_userspace_map(void *address, unsigned long pages, void **user_address, void **handle);
-extern void co_os_userspace_unmap(void *user_address, void *handle, unsigned long pages);
+extern co_rc_t co_os_userspace_map(void *address, unsigned int pages, void **user_address, void **handle);
+extern void co_os_userspace_unmap(void *user_address, void *handle, unsigned int pages);
 
 #endif
